Trial Selector and import.sql removal

+++ src/main/java/com/tekenable/model/BaseEntity.java
@@ -4,15 +4,15 @@ import javax.persistence.*;
 
 @MappedSuperclass
 public class BaseEntity {
-  protected Long id;
+  protected int id;
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  public Long getId() {
+  public int getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(int id) {
     this.id = id;
   }
 }
